fix: sum should be a double

+++ lab03_session02_openMP/main-noopti.c
@@ -196,7 +196,7 @@ double sigmoid(double input)
 void forward_pass(NeuralNetwork *network)
 {
 	for (int i = 0; i < HIDDEN_SIZE; ++i) {
-		int sum = 0;
+		double sum = 0;
 		for (int j = 0; j < INPUT_SIZE; ++j) {
 			sum += network->input[j] * network->weights_ih[j][i];
 		}
@@ -204,7 +204,7 @@ void forward_pass(NeuralNetwork *network)
 	}
 
 	for (int i = 0; i < OUTPUT_SIZE; ++i) {
-		int sum = 0;
+		double sum = 0;
 		for (int j = 0; j < HIDDEN_SIZE; ++j) {
 			sum += network->hidden[j] * network->weights_ho[j][i];
 		}
@@ -234,7 +234,7 @@ void backpropagation(NeuralNetwork *network, double target[OUTPUT_SIZE])
 		}
 	}
 	for (int i = 0; i < HIDDEN_SIZE; ++i) {
-		int sum = 0;
+		double sum = 0;
 		network->hidden[i] = 0;
 		for (int j = 0; j < OUTPUT_SIZE; ++j) {
 			sum += output_gradients[j] * network->weights_ho[i][j] *
